Cascade

+++ src/main/java/com/stayhomes/servlets/DeleteUserServlet.java
@@ -1,4 +1,4 @@
-package com.stayhome.servlets;
+package com.stayhomes.servlets;
 
 import com.stayhome.dao.DbCon;
 
@@ -10,6 +10,7 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 
 @WebServlet(name = "deleteUser", value = "/deleteUser")
 public class DeleteUserServlet extends HttpServlet {
@@ -19,17 +20,33 @@ public class DeleteUserServlet extends HttpServlet {
         int userId = Integer.parseInt(request.getParameter("id"));
 
         try {
-            // Delete the user from the database
+            // Check if the user's usertype is 1 (owner)
             Connection con = DbCon.getConn();
-            PreparedStatement pst = con.prepareStatement("DELETE FROM users WHERE id = ?");
-            pst.setInt(1, userId);
-            pst.executeUpdate();
+            PreparedStatement checkPst = con.prepareStatement("SELECT user_type FROM users WHERE id = ?");
+            checkPst.setInt(1, userId);
+            ResultSet rs = checkPst.executeQuery();
 
-            // Set the message parameter
-//           request.setAttribute("message", "User deleted successfully");
+            if (rs.next()) {
+                int userType = rs.getInt("user_type");
 
-            // Redirect to the home page with the message parameter
-            response.sendRedirect("index.jsp?message=User deleted successfully");
+                if (userType == 1) {
+                    // Update the flat_owner ID to NULL in flats table
+                    PreparedStatement updatePst = con.prepareStatement("UPDATE flats SET flat_owner = NULL WHERE flat_owner = ?");
+                    updatePst.setInt(1, userId);
+                    updatePst.executeUpdate();
+                }
+
+                // Delete the user from the database
+                PreparedStatement deletePst = con.prepareStatement("DELETE FROM users WHERE id = ?");
+                deletePst.setInt(1, userId);
+                deletePst.executeUpdate();
+
+                // Redirect to the home page with the message parameter
+                response.sendRedirect("index.jsp?message=User deleted successfully");
+            } else {
+                // User not found, display appropriate message or redirect to an error page
+                response.sendRedirect("index.jsp?message=User not found");
+            }
         } catch (Exception e) {
             e.printStackTrace();
         }
